+++ SIDI/purposedSteps/_build_e2e_android/config.ts
@@ -1,7 +1,6 @@
 import { IConfig } from '../../../src/models/SidiModel';
 import androidManifest from '../../../src/customChecks/AndroidManifest';
 import { GluegunToolbox } from 'gluegun';
-import { androidConfig, commonConfig, rnConfig } from '../configs';
 
 const config: IConfig = {
   mandatoryKeys: [
@@ -29,7 +28,7 @@ const config: IConfig = {
     '_e2e_upload_apps_artifacts',
     '_e2e_launch_browserstack_builds',
     '_e2e_zip_artifacts',
-  ]
+  ],
 };
 
 export default config;