+++ SIDI/purposedSteps/_promote_app/config.ts
@@ -1,9 +1,9 @@
 import { IConfig } from '../../../src/models/SidiModel';
-import { commonConfig, qualityConfig, rnConfig } from '../configs';
+import { commonConfig, flutterConfig, qualityConfig, rnConfig } from '../configs';
 
 const config: IConfig = {
   mandatoryKeys: [],
-  conditionalSteps: [...commonConfig, ...rnConfig, ...qualityConfig],
+  conditionalSteps: [...commonConfig, ...rnConfig, ...flutterConfig, ...qualityConfig],
   stepsFamily: ['_promote_app_step'],
 };