+++ SIDI/steps/_react_native_npm_library_install/bitrise/config.ts
@@ -1,7 +1,10 @@
 import { IConfig } from '../../../../src/models/SidiModel';
 
 const config: IConfig = {
-  mandatoryKeys: [{ key: 'SAMPLE_APP_PATH', globalValue: true, defaultValue: 'example' }],
+  mandatoryKeys: [
+    { key: 'SAMPLE_APP_PATH', globalValue: true, defaultValue: 'example' },
+    { key: 'NPM_VERSION', globalValue: true },
+  ],
 };
 
 export default config;